Split into files as in the package
