+++ types/onflow__flow-js-testing/src/contract.d.ts
@@ -1,0 +1,3 @@
+import { Address } from './account';
+
+export function getContractAddress(name: string): Promise<Address>;
