+++ types/onflow__flow-js-testing/src/emulator.d.ts
@@ -1,0 +1,15 @@
+export class Emulator {
+    start(options?: {
+        logging?: boolean;
+        flags?: string;
+        adminPort?: number;
+        restPort?: number;
+        grpcPort?: number;
+    }): Promise<boolean>;
+
+    stop(): Promise<boolean>;
+
+    setLogging(logging: boolean): void;
+}
+
+export const emulator: Emulator;
