+++ types/onflow__flow-js-testing/src/flow-token.d.ts
@@ -1,0 +1,6 @@
+import { Address } from './account';
+import { TransactionResponse } from './interaction';
+
+export function getFlowBalance(address: Address): Promise<string>;
+
+export function mintFlow(recipient: Address, amount: string): Promise<TransactionResponse>;
