+++ types/onflow__flow-js-testing/src/init.d.ts
@@ -1,0 +1,1 @@
+export function init(bastPath: string, options?: { pkey?: string }): Promise<boolean>;
