+++ types/onflow__flow-js-testing/src/manager.d.ts
@@ -1,0 +1,7 @@
+export function getBlockOffset(): Promise<string>;
+
+export function setBlockOffset(offset: string): Promise<string>;
+
+export function getTimestampOffset(): Promise<string>;
+
+export function setTimestampOffset(offset: string): Promise<string>;
