+++ types/onflow__flow-js-testing/src/templates.d.ts
@@ -1,0 +1,9 @@
+import { AddressMap } from './account';
+
+export function getTemplate(file: string, addressMap?: AddressMap, byAddress?: boolean): Promise<string>;
+
+export function getContractCode(props: { name: string; addressMap?: AddressMap }): Promise<string>;
+
+export function getTransactionCode(props: { name: string; addressMap?: AddressMap }): Promise<string>;
+
+export function getScriptCode(props: { name: string; addressMap?: AddressMap }): Promise<string>;
