+++ types/onflow__flow-js-testing/src/transformers.d.ts
@@ -1,0 +1,1 @@
+export type CadenceTransformer = (code: string) => string;
