+++ types/onflow__flow-js-testing/src/utils.d.ts
@@ -1,0 +1,1 @@
+export function isAddress(address: string): boolean;
